packages/video-ocr/pyproject.toml
[project]
name = "video-ocr"
version = "0.1.0"
description = "Extract OCR text and page-transition keyframes from screen recordings"
requires-python = ">=3.10"
dependencies = [
    "paddleocr>=2.8.0",
    "paddlepaddle>=2.6.0",
    "opencv-python>=4.9.0",
    "imagehash>=4.3.1",
    "pillow>=10.0.0",
    "numpy>=1.24.0,<2.0",
    "typer>=0.12.0",
]

[project.scripts]
video-ocr = "video_ocr.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_ocr"]
